+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Changed default log level to `INFO`
+
+
 0.2.16
 ------
 - Use `tracing-subscriber`'s `tracing-log` feature to unify log output